Moved all Port definitions in a separate file
