+++ discovery.go
@@ -36,19 +36,7 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-
-	"github.com/arduino/go-properties-orderedmap"
 )
-
-// Port is a descriptor for a board port
-type Port struct {
-	Address       string          `json:"address"`
-	AddressLabel  string          `json:"label,omitempty"`
-	Protocol      string          `json:"protocol,omitempty"`
-	ProtocolLabel string          `json:"protocolLabel,omitempty"`
-	Properties    *properties.Map `json:"properties,omitempty"`
-	HardwareID    string          `json:"hardwareId,omitempty"`
-}
 
 // Discovery is an interface that represents the business logic that
 // a pluggable discovery must implement. The communication protocol
